refactor: move OIDC related classes under sm package
