+++ migration/identity-migration/src/main/java/io/camunda/migration/identity/config/sm/OidcProperties.java
@@ -5,7 +5,7 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.migration.identity.config.oidc;
+package io.camunda.migration.identity.config.sm;
 
 public class OidcProperties {
