+++ migration/identity-migration/src/main/java/io/camunda/migration/identity/config/sm/SMKeycloakMigrationHandlerConfig.java
@@ -9,7 +9,7 @@ package io.camunda.migration.identity.config.sm;
 
 import io.camunda.migration.identity.client.ManagementIdentityClient;
 import io.camunda.migration.identity.config.IdentityMigrationProperties;
-import io.camunda.migration.identity.handler.RoleMigrationHandler;
+import io.camunda.migration.identity.handler.sm.RoleMigrationHandler;
 import io.camunda.migration.identity.handler.sm.AuthorizationMigrationHandler;
 import io.camunda.migration.identity.handler.sm.ClientMigrationHandler;
 import io.camunda.migration.identity.handler.sm.GroupMigrationHandler;
@@ -19,6 +19,7 @@ import io.camunda.security.auth.CamundaAuthentication;
 import io.camunda.service.AuthorizationServices;
 import io.camunda.service.GroupServices;
 import io.camunda.service.RoleServices;
+import io.camunda.service.TenantServices;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.core.annotation.Order;
@@ -87,7 +88,7 @@ public class SMKeycloakMigrationHandlerConfig {
   @Order(6)
   public TenantMigrationHandler tenantMigrationHandler(
       final ManagementIdentityClient managementIdentityClient,
-      final io.camunda.service.TenantServices tenantService,
+      final TenantServices tenantService,
       final CamundaAuthentication camundaAuthentication) {
     return new TenantMigrationHandler(
         managementIdentityClient, tenantService, camundaAuthentication);
